[RFC] Only store persistable state on unmount
Currently we write to persistable state on each change to any persistable state including the current query and variables as well as some the resizable elements of the UI. This suggests replacing that with only saving persistable state when the component is about to unmount.

This fixes a problem where opening multiple instances of GraphiQL at the same time can result in one query appearing across all views. Now, only the last window closed will save the query. This also fixes a frustration where accidentally closing a tab and then editing another instance of GraphiQL renders the accidentally closed tab harder to recover.

This also fixes a performance issue where setting into storage can be slightly expensive and we were performing this on every keystroke.

The primary drawback is the potential for lost state in the case of an unexpected browser cache, since this no longer saves to persistable state optimisitically. I believe that's worth the gains and a minor issue relative to the accidental tab-closing issue.