Replace inline documentation with reference to README
Instead of three sources of truth (README, code docs, actual code),
let's cut back to two (README, actual code).

An argument could be made that it would be easier to keep code docs and
actual code in sync than to keep the README and the code in sync, but
experience shows that even when documentation is adjacent to code it can
easily get out of date: so, let's go with the README where we can get
this stuff in front of as many eyeballs as possible and extract maximum
value from it.