Run introspection again if fetcher changes.
Closes #223

Right now, if a new fetching function is provided, GraphiQL doesn't refetch the schema with it. That can result in different schema running queries and used within the tool for insights.

This change causes the new fetcher to be called with introspection to get a new schema.

If GraphiQL is rerendered by the fetcher function is the same, then nothing will happen.

BREAKING: Note that if you provide a fetching function that is generated on every render, such as a non-autobound function in a parent React component, this will cause it to re-fetch the schema every time it's rerendered. To avoid this, it's important to ensure the fetching function changes only when intentionally using new behavior (such as a new endpoint to fetch from).