Default toolbar with prettify button.
Currently there is a hard-coded prettify button in the toolbar, however not all clients of GraphQL wish to have that button, or to have that button in that particular position. This PR replaces hard-coding with a default toolbar that contains the prettify button.

This means if you pass a `<GraphiQL.Toolbar>` child to `<GraphiQL>` that the Prettify button will no longer appear. If you want the Prettify button, you can copy the existing source. Here is an example of adding a Toolbar with a Prettify button:

```js
const { parse, print } = require('graphql');

class CustomGraphiQL extends React.Component {
  render() {
    return (
      <GraphiQL ref={c => this.graphiql = c} ...>
        <GraphiQL.Toolbar>
          <GraphiQL.Button title="Prettify" onClick={this.handlePrettify}>
        </GraphiQL.Toolbar>
      </GraphiQL>
    );
  }

  handlePrettify = () => {
    const editor = this.graphiql.getQueryEditor();
    editor.setValue(print(parse(editor.getValue())));
  }
}
```