Guard against unset `state.rule` (#112)
Some kinds of bad JSON were causing an exception here. Example:

  {}garbage

would cause `state.rule.length` to blow up with:

  Uncaught TypeError: Cannot read property 'length' of undefined

This applies the same fix that already exists in
graphql-language-service (and is another reason why this common code
should be factored out into a package that both codemirror-graphql and
graphql-language-service can depend on; will have PRs up for that soon).