Extract types into a separate package
This is step one, which creates all the support necessary to publish a
new package, "graphql-language-service-types", but doesn't depend on it
yet. Once that is published, we'll proceed to step two, which is
depending on the published types, and step three, which is figuring out
whether we should dive in and embrace Lerna, or do it manually for a
while (either way, need to figure out whether we want to run this
independently in CI, at what level we want to set up Flow etc).

This commit is a straight copy of "types/Types.js" to "src/index.js" in
the new package's directory, plus a new "package.json" and other
supporting paraphernalia like a script for creating "*.js.flow" files in
the distributed bundle.