Depend on graphql-language-service-types
Depend on the extracted package, instead of the local copies of the
files. To develop against the local copy you have to use `npm link` or
`yarn link` for now, but I'll try out Lerna in a bit which will smooth
over some of the rough edges with the multi-package development process.