Get tests passing by breaking dependency on other packages
Don't depend on the server package (which in turn depends on the config
package). This allows us to avoid an otherwise very hard to disentangle
circular dependency.

On the down side, this means the test is no longer the end-to-end
integration test that it was before, in some sense, but the argument
could be made that it never should have been (it's a unit test). Also,
note that it wasn't running before because of the "-spec.js" extension
(we only run files matching "-test.js") so its value was limited anyway.

Let's replace this with an actual integration test in the top-level
package later on.