Remove @flow directives from tests
We will likely want to typecheck these later, but for now I just want to
quickly reorganize the files that we have, so I'm going to strip the
annotation.

Note that I could also have modified the .flowconfig, but I prefer not
to do a blanket exclusion of all __tests__, so that we can leave open a
doorway for incremental adoption of Flow on a file-by-file basis in the
future.