Tweak linting to work with prettier
Prettier will try to wrap lines to 80 by default, but sometimes it can't
do so. In those cases, we don't want eslint to yell at us, so let's up
the max-len to give us some breathing room. This fixes 4 lint warnings
in the graphql-language-service-interface package.

The other tweak is rejigging the code to use an explicit `return`,
because otherwise eslint complains about "Arrow function used
ambiguously with a conditional expression" (no-confusing-arrow).