Reset initial size of main and doc panels when double-clicking their resize area
It is fairly common to see tools with this capability, so I mechanically tried to double-click the resize area to reset the main section to 50%/50% before realizing I was not on a desktop tool :)

After adding this to the main section, I figured docs section could benefit from it as well. Because of how toggling variables section works (single click), it is not possible to add this to the variables section at the moment. I did not have enough time to investigate more as this would require a bigger UI change, which I would not do without starting a discussion anyway, so it is better left to a future work. This already adds value as it is, I think.