fix: remove statement pending discussion
I made the choice of adding this to the readme w/o consulting colleauges

My co-maintainers are being forced to take sides and this is unfair

The technical steering commitee will review this decision, & decide
What best reflects our interests together
First and foremost, they will decide on a policy to make sure a rouge
maintainer action like this cannot happen again without input from the TSC.

I apologize for making this choice but not what I said, and will continue
supporting an end to the apartheid in my own personal venues.

Please leave my co-maintainers and colleauges be.