Merge pull request from GHSA-x4r7-m2q9-69c8
This change fixes an XSS vulnerability that has been present since the first
commit of GraphiQL when used with an untrusted GraphQL server. It applies three
strategies to fix via defense-in-depth: HTML escaping a string in a contexts
that uses innerHTML; validating incoming schemas with graphql-js validateSchema;
and updating markdown-it to the latest version.

See docs/security/2021-introspection-schema-xss.md for more details.

Co-authored-by: Rikki <rikki.schulte@gmail.com>