Add Cristina Shaver as co-maintainer (#2848)
Cristina began helping with `monaco-graphql` and `graphiql` in late 2019 or 2020, providing some of the implementation that's in both `@graphiql/react` and `monaco-graphql` today. She has also helped with issue triage in the past.

She would like to help us advance`monaco-graphql` and `graphiql` and `@graphiql/react`. Her eye for a11y and both DX and UX will help so much!